+++ components/Chips/examples/supplemental/ChipsExamplesSupplemental.m
@@ -40,7 +40,7 @@ static UIButton *DeleteButton() {
   image = [image imageWithRenderingMode:UIImageRenderingModeAlwaysTemplate];
 
   UIButton *button = [[UIButton alloc] initWithFrame:CGRectZero];
-  button.tintColor = [UIColor colorWithWhite:0 alpha:0.7f];
+  button.tintColor = [UIColor colorWithWhite:0 alpha:(CGFloat)0.7];
   [button setImage:image forState:UIControlStateNormal];
 
   return button;
