+++ components/Ink/src/private/MDCInkLayer.m
@@ -23,15 +23,15 @@ static NSString *const MDCInkLayerInitialRadiusKey = @"MDCInkLayerInitialRadiusK
 static NSString *const MDCInkLayerMaxRippleRadiusKey = @"MDCInkLayerMaxRippleRadiusKey";
 static NSString *const MDCInkLayerInkColorKey = @"MDCInkLayerInkColorKey";
 
-static const CGFloat MDCInkLayerCommonDuration = 0.083f;
-static const CGFloat MDCInkLayerEndFadeOutDuration = 0.15f;
-static const CGFloat MDCInkLayerStartScalePositionDuration = 0.333f;
-static const CGFloat MDCInkLayerStartFadeHalfDuration = 0.167f;
-static const CGFloat MDCInkLayerStartFadeHalfBeginTimeFadeOutDuration = 0.25f;
+static const CGFloat MDCInkLayerCommonDuration = (CGFloat)0.083;
+static const CGFloat MDCInkLayerEndFadeOutDuration = (CGFloat)0.15;
+static const CGFloat MDCInkLayerStartScalePositionDuration = (CGFloat)0.333;
+static const CGFloat MDCInkLayerStartFadeHalfDuration = (CGFloat)0.167;
+static const CGFloat MDCInkLayerStartFadeHalfBeginTimeFadeOutDuration = (CGFloat)0.25;
 
-static const CGFloat MDCInkLayerScaleStartMin = 0.2f;
-static const CGFloat MDCInkLayerScaleStartMax = 0.6f;
-static const CGFloat MDCInkLayerScaleDivisor = 300.f;
+static const CGFloat MDCInkLayerScaleStartMin = (CGFloat)0.2;
+static const CGFloat MDCInkLayerScaleStartMax = (CGFloat)0.6;
+static const CGFloat MDCInkLayerScaleDivisor = 300;
 
 static NSString *const MDCInkLayerOpacityString = @"opacity";
 static NSString *const MDCInkLayerPositionString = @"position";
@@ -129,8 +129,8 @@ static NSString *const MDCInkLayerScaleString = @"transform.scale";
 }
 
 - (void)setRadiiWithRect:(CGRect)rect {
-  self.initialRadius = (CGFloat)(MDCHypot(CGRectGetHeight(rect), CGRectGetWidth(rect)) / 2 * 0.6f);
-  self.finalRadius = (CGFloat)(MDCHypot(CGRectGetHeight(rect), CGRectGetWidth(rect)) / 2 + 10.f);
+  self.initialRadius = (CGFloat)(MDCHypot(CGRectGetHeight(rect), CGRectGetWidth(rect)) / 2 * (CGFloat)0.6);
+  self.finalRadius = (CGFloat)(MDCHypot(CGRectGetHeight(rect), CGRectGetWidth(rect)) / 2 + 10);
 }
 
 - (void)startAnimationAtPoint:(CGPoint)point {
@@ -158,7 +158,7 @@ static NSString *const MDCInkLayerScaleString = @"transform.scale";
     _startAnimationActive = YES;
 
     CAMediaTimingFunction *materialTimingFunction =
-        [[CAMediaTimingFunction alloc] initWithControlPoints:0.4f:0:0.2f:1.f];
+        [[CAMediaTimingFunction alloc] initWithControlPoints:(CGFloat)0.4:0:(CGFloat)0.2:1];
 
     CGFloat scaleStart =
         MIN(CGRectGetWidth(self.bounds), CGRectGetHeight(self.bounds)) / MDCInkLayerScaleDivisor;
@@ -171,7 +171,7 @@ static NSString *const MDCInkLayerScaleString = @"transform.scale";
     CABasicAnimation *scaleAnim = [[CABasicAnimation alloc] init];
     scaleAnim.keyPath = MDCInkLayerScaleString;
     scaleAnim.fromValue = @(scaleStart);
-    scaleAnim.toValue = @1.0f;
+    scaleAnim.toValue = @1;
     scaleAnim.duration = MDCInkLayerStartScalePositionDuration;
     scaleAnim.beginTime = MDCInkLayerCommonDuration;
     scaleAnim.timingFunction = materialTimingFunction;
@@ -188,8 +188,8 @@ static NSString *const MDCInkLayerScaleString = @"transform.scale";
     CAKeyframeAnimation *positionAnim = [[CAKeyframeAnimation alloc] init];
     positionAnim.keyPath = MDCInkLayerPositionString;
     positionAnim.path = centerPath.CGPath;
-    positionAnim.keyTimes = @[ @0, @1.0f ];
-    positionAnim.values = @[ @0, @1.0f ];
+    positionAnim.keyTimes = @[ @0, @1 ];
+    positionAnim.values = @[ @0, @1 ];
     positionAnim.duration = MDCInkLayerStartScalePositionDuration;
     positionAnim.beginTime = MDCInkLayerCommonDuration;
     positionAnim.timingFunction = materialTimingFunction;
@@ -199,7 +199,7 @@ static NSString *const MDCInkLayerScaleString = @"transform.scale";
     CABasicAnimation *fadeInAnim = [[CABasicAnimation alloc] init];
     fadeInAnim.keyPath = MDCInkLayerOpacityString;
     fadeInAnim.fromValue = @0;
-    fadeInAnim.toValue = @1.0f;
+    fadeInAnim.toValue = @1;
     fadeInAnim.duration = MDCInkLayerCommonDuration;
     fadeInAnim.beginTime = MDCInkLayerCommonDuration;
     fadeInAnim.timingFunction = [CAMediaTimingFunction functionWithName:kCAMediaTimingFunctionLinear];
@@ -234,7 +234,7 @@ static NSString *const MDCInkLayerScaleString = @"transform.scale";
   CGFloat currOpacity = self.presentationLayer.opacity;
   CGFloat updatedOpacity = 0;
   if (viewContainsPoint) {
-    updatedOpacity = 1.0f;
+    updatedOpacity = 1;
   }
 
   CABasicAnimation *changeAnim = [[CABasicAnimation alloc] init];
@@ -259,7 +259,7 @@ static NSString *const MDCInkLayerScaleString = @"transform.scale";
     self.endAnimationDelay = MDCInkLayerStartFadeHalfBeginTimeFadeOutDuration;
   }
 
-  CGFloat opacity = 1.0f;
+  CGFloat opacity = 1;
   BOOL viewContainsPoint = CGRectContainsPoint(self.bounds, point) ? YES : NO;
   if (!viewContainsPoint) {
     opacity = 0;
